requirements.txt
bcrypt==4.0.1
cffi==1.15.1
cryptography==38.0.3
nmap==0.0.1
paramiko==2.12.0
pycparser==2.21
pydig==0.4.0
PyNaCl==1.5.0
python-nmap==0.7.1
python3-nmap==1.6.0
scapy==2.4.5
simplejson==3.17.6
six==1.16.0
